rename context api folder
