+++ src/core/services/notification/notification.stories.ts
@@ -1,12 +1,12 @@
 import '../../../backoffice/components/backoffice-notification-container.element';
-import '../../context/provide/context-provider.element';
+import '../../context-api/provide/context-provider.element';
 import './layouts/default';
 
 import { Meta, Story } from '@storybook/web-components';
 import { html, LitElement } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
-import { UmbContextConsumerMixin } from '../../context';
+import { UmbContextConsumerMixin } from '../../context-api';
 
 import type { UmbNotificationDefaultData } from './layouts/default';
 import { UmbNotificationColor, UmbNotificationOptions, UmbNotificationService } from '.';
