+++ tsconfig.json
@@ -26,7 +26,7 @@
     "paths": {
       "@umbraco-cms/models": ["src/core/models"],
       "@umbraco-cms/backend-api": ["src/core/backend-api"],
-      "@umbraco-cms/context-api": ["src/core/context"],
+      "@umbraco-cms/context-api": ["src/core/context-api"],
       "@umbraco-cms/extensions-api": ["src/core/extensions-api"],
       "@umbraco-cms/observable-api": ["src/core/observable-api"],
       "@umbraco-cms/utils": ["src/core/helpers"],
